requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.1
certifi==2024.8.30
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
diff-match-patch==20241021
Django==5.1.1
django-cors-headers==4.4.0
django-environ==0.11.2
django-import-export==4.2.0
django-ratelimit==4.1.0
django-redis==5.4.0
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.7.0
drf-yasg==1.21.10
executing==2.2.0
Flask==3.1.0
fonttools==4.55.3
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kombu==5.5.3
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mysql-connector-python==9.1.0
mysqlclient==2.2.4
nest-asyncio==1.6.0
numpy==2.1.2
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.9.0
pymongo==4.11.3
pyparsing==3.2.0
PySocks==1.7.1
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.2.1
selenium==4.26.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
stack-data==0.6.3
tablib==3.7.0
tornado==6.4.2
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
